src/python/ai_querier/requirements.txt
pymongo==4.5.0
redis==4.5.5
google-generativeai